requirements.txt
apscheduler>=3.10.4
beautifulsoup4>=4.12.3
dnspython>=2.7.0
flask>=3.1.0
google-auth-oauthlib>=1.2.0
google-genai>=1.9.0
gspread>=6.1.4
gunicorn>=23.0.0
jinja2>=3.1.5
openpyxl>=3.1.5
pandas>=2.2.3
python-dotenv>=1.0.1
python-telegram-bot>=21.10
requests>=2.32.3
rich>=13.9.4
tqdm>=4.67.1
